+++ routes/api.php
@@ -16,9 +16,9 @@ Route::get('/', 'HomeController');
 
 Route::group(['middleware' => ['auth:api']], function () {
     Route::get('user', 'Settings\ProfileController@index');
-    Route::get('asset/list', 'ListController@index');
-    Route::apiResource('asset', 'AssetController');
     Route::group(['namespace' => 'V1'], function () {
+        Route::get('asset/list', 'ActiveListAssetController@index');
+        Route::apiResource('asset', 'AssetController');
         Route::apiResource('reservation', 'ReservationController')->except('update');
         Route::apiResource('reserved', 'ReservedController')
             ->only(['index', 'update'])
